Give support for --cpu-arch build option when on MSVC

+++ src/kryptos_feal.c
@@ -24,7 +24,7 @@
 #define kryptos_feal_get_byte_from_u16(w, n) ( ( (w) >> (8 - ((n) << 3)) ) & 0xff )
 
 struct kryptos_feal_subkeys {
- int rounds;
+ size_t rounds;
  kryptos_u16_t K[KRYPTOS_FEAL_MAX];
 };
